Moved dot plot series into subfolder.
