+++ js/masters/modules/dotplot.src.js
@@ -10,4 +10,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Series/DotplotSeries.js';
+import '../../Series/DotPlot/DotPlotSeries.js';
